requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
Werkzeug==2.3.7
